# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: 3ds2-demo Payment API
  version: 0.1.0
  description: |
    EMV 3-D Secure 2.x 統合学習デモの Payment API 契約定義。
    Stripe Payment Intents API をベースに、3DS2 challenge フローを
    明示的にハンドリングする backend の REST 仕様。

    ## 関連ドキュメント
    - 3DS2 仕様概観: `docs/3ds2-overview-ja.md`
    - State Machine 詳細: `docs/architecture.md`

  contact:
    name: 3ds2-demo
    url: https://github.com/izdev0000/3ds2-demo
  license:
    name: MIT
    url: https://github.com/izdev0000/3ds2-demo/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: ローカル開発環境（docker compose）
...
paths:
  /api/orders:
    post:
      tags: [orders]
      summary: 仮注文を作成
      description: |
        決済前段階の Order を `pending` 状態で作成する。
        Order は業務上の注文単位であり、決済（Transaction）は 1 Order に対し N 件
        紐づく（失敗で別カード再決済するケースに対応）。

        Order.status は webhook 経由でのみ `paid` に遷移する。
        frontend からの confirm 成功 response や return_url 到達では遷移しない
        （docs/design/error-handling.md §8.4 参照）。
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              single_item:
                summary: 商品 1 点の注文
                value:
                  currency: jpy
                  items:
                    - name: T-shirt (S)
                      quantity: 1
                      unit_price: 1000
              multi_item:
                summary: 商品複数点の注文
                value:
                  currency: jpy
                  items:
                    - name: T-shirt (S)
                      quantity: 2
                      unit_price: 1000
                    - name: Sticker pack
                      quantity: 1
                      unit_price: 500
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Order の現在状態を取得
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    OrderIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^ord_[A-Za-z0-9]+$'
      description: 内部 order ID（`ord_` プレフィックス + ULID 等）
      example: ord_01HYZ800
  schemas:
    CreateOrderRequest:
      type: object
      required: [currency, items]
      properties:
        currency:
          type: string
          minLength: 3
          maxLength: 3
          pattern: '^[a-z]{3}$'
          description: ISO 4217 通貨コード（小文字）
          example: jpy
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemInput'
        metadata:
          type: object
          additionalProperties: true
          description: 任意の業務メタデータ
    OrderItemInput:
      type: object
      required: [name, quantity, unit_price]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: T-shirt (S)
        quantity:
          type: integer
          minimum: 1
          example: 1
        unit_price:
          type: integer
          minimum: 0
          description: 通貨の最小単位（JPY なら円、USD ならセント）
          example: 1000
    OrderResponse:
      type: object
      required: [id, status, amount, currency, items, created_at, updated_at]
      properties:
        id:
          type: string
          description: 内部 order ID
          example: ord_01HYZ800
        status:
          $ref: '#/components/schemas/OrderStatus'
        amount:
          type: integer
          description: items の合計金額（quantity * unit_price の総和）
          example: 1000
        currency:
          type: string
          example: jpy
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        metadata:
          type:
            - object
            - 'null'
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      allOf:
        - $ref: '#/components/schemas/OrderItemInput'
        - type: object
          required: [id, subtotal]
          properties:
            id:
              type: string
              description: 内部 order item ID
              example: oit_01HYZ800
            subtotal:
              type: integer
              description: quantity * unit_price
              example: 1000
    OrderStatus:
      type: string
      description: |
        Order の業務状態。webhook 経由でのみ `paid` に遷移する。
        | status | 意味 |
        | --- | --- |
        | pending | 仮注文（決済未確定） |
        | paid | 決済成功・注文確定 |
        | canceled | 注文キャンセル（決済前 or 全 Transaction が canceled） |
        | refunded | 返金済み |
      enum:
        - pending
        - paid
        - canceled
        - refunded
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: アプリケーション固有のエラーコード
          enum:
            - invalid_amount
            - invalid_currency
            - invalid_payment_method
            - validation_failed
            - payment_not_found
            - payment_already_confirmed
            - order_not_found
            - order_already_paid
            - order_already_canceled
            - signature_verification_failed
            - stripe_api_error
            - internal_error
          example: invalid_amount
        message:
          type: string
          description: 人間可読なエラーメッセージ
          example: amount must be at least 50
        details:
          type: object
          additionalProperties: true
          description: エラー固有の追加情報
  responses:
...
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: リクエストは valid だが状態的に処理できない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...